.
insert essentially done afaict. There may be an off by one in node splitting
condition (N >= BT_DAT_MAXKEYS - 1), particularly for insertion into leaf nodes.

We do kind of eagerly split in branches because even if the child is dirty (no
insertion required) we still split if width is at N-1

So, for instance:
- (1) insert into path P (fully dirty) : for some branch node, B, N = BT_DAT_MAXKEYS - 2
- (2) insert into path P' : for branch node, B, N = BT_DAT_MAXKEYS - 1
- (3) insert again into path P

(3) will cause a split in branch node, B, even though it's unnecessary (since
the child is already dirty and therefore no CoW and therefore no insertion).

---

going to attempt some of delete and work on freelists for now