.
all of the below should still be considered wip and untested:

- flist and mlist creation done on initial reading of btree state from pma file

- appropriate insertion of ranges into mlist on bt_malloc done

- _bt_falloc mostly fine for now (add 2M assertion?) but will need to have its
logic adjusted when the persistent file is implemented to appropriately extend
the end of the file and mind the striping

working on bt delete next and reviewing insertion logic. After which, we should
be able to move on to persistence implementation. May want to start testing the
code before that and debugging the inevitable bugs that are present here