.
not the current status, just a checkpoint for now that compiles. Going to
continue work a bit longer and push an overriding commit with the full nlist
implementation, (nalloc, nlist_new (ez), nlist_read (more challenging though not
terrible - about equivalent to mlist_read though completely different
logic)). Deletion coalescing depends of course on the node freelist being a real
free list and not just a bump pointer. Should be back to work on the core
deletion coalescing logic tomorrow. first pass is finished

one other thing that's missing: we have allocation routines for mlist and flist,
but not freeing routines. These need to be implemented.