.
deletion coalescing mostly finished. still need to add logic to invoke this in
the insert routine. insert routine may need some refactoring. why?:

- currently it directly handles ensuring a node is either dirty or
cowed. Additionally ensures that if a node will be made full by the insertion
that it is split.

- deletion coalescing will run prior to any insertion. The path of the deletion
coalescing is the same. Therefore, we should handle the above
requirements (dirtying, cowing, (and /maybe/ splitting - not sure yet)) in the
deletion coalescing path.

---

a minor bug fix or two also included in this commit. For instance, dependency on
state->depth was removed and now the metapage is the only thing tracking
depth. This is more robust in that we don't have to ensure the two separate
values are in sync